chore: unify protocol types on protobuf, remove serde_json, add lifegraphd binary
- Remove lifegraph-agent (experimental daemon) and systemd service
- Eliminate CSE-v0 CBOR canonicalization — use protobuf wire encoding directly
- Delete lifegraph-core/src/cbor.rs, canonical.rs, metadata.rs, convert.rs (~4600 lines)
- Core types are now protobuf-generated structs re-exported from lifegraph-proto
- No more From<&Core> for Proto conversion layer
- Remove serde_json from all crates; cargo-deny bans it workspace-wide
- Replace BLAKE3 with SHA-256 for all content hashing (universal hardware support)
- Stream crate uses core EventEnvelope directly (no duplicate struct)
- Add lifegraphd binary with init/run/status commands
- Init detects TPM/YubiKey hardware; --software flag for dev-only insecure keys
- No separate .key file — config contains signer type and key reference
- Add deny.toml banning serde_json as direct dependency
- Update spec §17: canonical encoding is protobuf wire format

+++ rust/crates/lifegraph-core/src/util.rs
@@ -2,15 +2,15 @@ use time::format_description::well_known::Rfc3339;
 use time::OffsetDateTime;
 use unicode_normalization::UnicodeNormalization;
 
-use crate::protocol::Timestamp;
+use prost_types::Timestamp;
 
 pub fn parse_rfc3339(value: &str) -> Result<OffsetDateTime, time::error::Parse> {
     OffsetDateTime::parse(value, &Rfc3339)
 }
 
-pub fn timestamp_parts(value: &str) -> Result<(i64, i64), time::error::Parse> {
+pub fn timestamp_parts(value: &str) -> Result<(i64, i32), time::error::Parse> {
     let t = parse_rfc3339(value)?;
-    Ok((t.unix_timestamp(), t.nanosecond() as i64))
+    Ok((t.unix_timestamp(), t.nanosecond() as i32))
 }
 
 pub fn parse_timestamp_value(value: &str) -> Result<Timestamp, time::error::Parse> {
